Avoid triggering ubsan unsigned-integer-overflow
This isn't undefined, but it's suspicious in most situations so
useful to be able to enable to find bugs.

+++ src/commands.c
@@ -2480,7 +2480,7 @@ cmd_require(void)
    skipblanks();
    get_pos(&fp);
    while (1) {
-      int diff = *ver++ - read_uint();
+      int diff = (int)*ver++ - (int)read_uint();
       if (diff > 0) break;
       if (diff < 0) {
 	 size_t i, len;
